Refactored route into multiple files
